% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_data.R
\name{dskin.AUC.analyze}
\alias{dskin.AUC.analyze}
\title{Calculate Compartment AUC}
\usage{
dskin.AUC.analyze(parameter, r.result = NULL, path = NULL,
  time.unit = "h", mass.unit = "mg", scale.with.cross.section = FALSE)
}
\arguments{
\item{parameter}{The DSkin Parameter Pack.}

\item{r.result}{DSkin result data structure ('collect.data') by the R-version of DSkin. If NULL, data is read from produced files
defined by the Parameter Pack.}

\item{path}{The path to the mass-over-time files (NULL for wd).}

\item{time.unit}{The unit of time used for the overview [sec,min,h].}

\item{mass.unit}{The mass unit for all layer compartments [ng,ug,mg].}

\item{scale.with.cross.section}{Logical value to indicate if layer cross-section is taken into account for
volume calculation of the different skin layers.}
}
\value{
Data.frame with compartmental names as column names and metrics as row names.
The following attributes are attached to the data.frame:
\describe{
 \item{mass.unit}{The unit used for masses}
 \item{time.unit}{The unit used for times}
 \item{vol.unit}{The unit used for volumes. Currently always ml, but might change for future releases.}
}
}
\description{
Calculates the Area under the Curve (AUC) in units of mass.unit * time.unit / ml for every logged compartment.
AUC.last calculation is only performed if pharamacokinietics is enabled in the sink compartment and
assumes no transport from skin layers to the sink compartment at the last simulated time point.
}
\seealso{
\code{\link{dskin.parameter}}

\code{\link{dskin.simulate}}
}
