% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_helper.R
\name{dskin.c.init}
\alias{dskin.c.init}
\title{Calulcate Initial Concentration}
\usage{
dskin.c.init(dose.in.mg, height.um, A.sqcm)
}
\arguments{
\item{dose.in.mg}{The loading dose in mg.}

\item{height.um}{The compartment height in micrometer.}

\item{A.sqcm}{The vehicle application area in \eqn{cm^{2}}{cm^2}}
}
\value{
The homogeneous concentration inside the compartment in mg/ml.
}
\description{
Calculates an initial concentration from compartment dimensions and loading dose assuming a homogenous
distribution of mass inside the compartment.
}
\examples{
# Initial concentration of 10 mg inside a compartment of 100 micrometer x 15 cm^2
dskin.c.init(10, 100, 15)
}
