% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.compartment.names}
\alias{dskin.compartment.names}
\title{Compartment Names}
\usage{
dskin.compartment.names(parameter)
}
\arguments{
\item{parameter}{The DSkin parameter pack.}
}
\value{
Vector of compartment names defined in the Parameter Pack including vehicle and sink.
}
\description{
Returns a list of compartment names from an dskin parameter pack.
}
\seealso{
\code{\link{dskin.parameter}}
}
