% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.conc.filenames}
\alias{dskin.conc.filenames}
\title{List of Compartment Filenames}
\usage{
dskin.conc.filenames(parameter, path = NULL)
}
\arguments{
\item{parameter}{The DSkin parameter pack.}

\item{path}{The path to CDP files (NULL for wd).}
}
\value{
A data frame of file names associated with CDPs with named columns (layer names).
}
\description{
Returns a list of file names associated with concentration-depth profiles (CDP) defined by a DSkin Parameter Pack.
}
\examples{
p <- dskin.parameter("project_1", 2, c("SC", "DSL"), "Patch", "Blood")
file.names <- dskin.conc.filenames(p)
}
\seealso{
\code{\link{dskin.parameter}}
}
