% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_helper.R
\name{dskin.convert.masses}
\alias{dskin.convert.masses}
\title{Convert masses}
\usage{
dskin.convert.masses(val, val_unit, target_unit)
}
\arguments{
\item{val}{The value [int or real].}

\item{val_unit}{The unit of the value [mg/ug/ng].}

\item{target_unit}{The unit the value should be converted into [mg/ug/ng].}
}
\value{
Converted masses (single value or compatible data structure)
}
\description{
Converts one or more values of a certain mass unit into another mass unit.
}
\examples{
# convert 10 ng to mg
dskin.convert.masses(10, "ng", "mg")
}
