% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_helper.R
\name{dskin.convert.times}
\alias{dskin.convert.times}
\title{Convert times}
\usage{
dskin.convert.times(val, val_unit, target_unit)
}
\arguments{
\item{val}{The value [int or real].}

\item{val_unit}{The unit of the value [sec/min/h].}

\item{target_unit}{The unit the value should be converted into [sec/min/h].}
}
\value{
Converted times (single value or compatible data structure)
}
\description{
Converts one or more values of a certain time unit into another time unit.
}
\examples{
# convert 10 sec to min
dskin.convert.times(10, "sec", "min")
}
