% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.fromJsonStr}
\alias{dskin.fromJsonStr}
\title{DSkin JSON to Parameter Pack}
\usage{
dskin.fromJsonStr(json.string)
}
\arguments{
\item{json.string}{The DSkin JSON string.}
}
\value{
The DSkin parameter pack.
}
\description{
Converts a DSkin JSON configuration string into a DSkin pramater pack.
\strong{Warning:} The Parameter Pack is not validated.
}
\seealso{
\code{\link{dskin.parameter}}
}
