% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_data.R
\name{dskin.mass.analyze}
\alias{dskin.mass.analyze}
\title{Compartmental Mass Analysis}
\usage{
dskin.mass.analyze(parameter, r.result = NULL, path = NULL,
  time.unit = "h", mass.unit = "mg")
}
\arguments{
\item{parameter}{The DSkin Parameter Pack.}

\item{r.result}{DSkin result data structure ('collect.data') by the R-version of DSkin. If NULL, data is read from produced files
defined by the Parameter Pack.}

\item{path}{The path to the mass-over-time files (NULL for wd).}

\item{time.unit}{The unit of time used for the overview [sec,min,h].}

\item{mass.unit}{The mass unit for all layer compartments [ng,ug,mg].}
}
\value{
Data.frame with compartmental names as column names and metrics as row names.
The following attributes are attached to the data.frame:
\describe{
 \item{mass.unit}{The unit used for masses}
 \item{time.unit}{The unit used for times}
}
}
\description{
Calculates maximum mass, time to maximum mass, minimum mass and time to minimum mass for all logged compartments.
}
\seealso{
\code{\link{dskin.parameter}}

\code{\link{dskin.simulate}}
}
