% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.mass.filenames}
\alias{dskin.mass.filenames}
\title{List of Compartment Filenames}
\usage{
dskin.mass.filenames(parameter, path = NULL)
}
\arguments{
\item{parameter}{The DSkin parameter pack.}

\item{path}{The path to mass-over-time files (NULL for wd).}
}
\value{
A data frame of file names associated with mass-over-time profiles with named columns (layer names).
}
\description{
Returns a list of file names associated with mass-over-time profiles defined by a DSkin Parameter Pack.
}
\examples{
p <- dskin.parameter("project_1", 2, c("SC", "DSL"), "Patch", "Blood")
file.names <- dskin.mass.filenames(p)
}
\seealso{
\code{\link{dskin.parameter}}
}
