% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_plot.R
\name{dskin.movie.masses}
\alias{dskin.movie.masses}
\title{Movie Creation for Mass-over-time profiles}
\usage{
dskin.movie.masses(parameter, r.result = NULL, remove.tmp.images = TRUE,
  path = NULL, max.cols = 3, screen.play = 10, frame.rate = 25,
  resolution = c(1280, 720), time.unit = "h", layer.unit = "mg",
  sink.unit = "ng", time.label.pos = "top", mass.range = c(0, NULL),
  time.range = c(0, NULL), ...)
}
\arguments{
\item{parameter}{The DSkin parameter pack.}

\item{r.result}{DSkin result data structure ('collect.data') by the R-version of DSkin. If NULL data is read from files defined by the Parameter Pack.}

\item{remove.tmp.images}{Logical parameter to indicate whether produced temporary images should be deleted after processing.}

\item{path}{The path to the mass-over-time files (NULL for wd).}

\item{max.cols}{The maximum number of columns in the plot.}

\item{screen.play}{The screen play time of the produced movie in seconds.}

\item{frame.rate}{The appoximate frame rate of the produced movie.}

\item{resolution}{The resolution of the movie in c(width, height).}

\item{time.unit}{The unit of time [sec,min,h].}

\item{layer.unit}{The unit of the y-axis [ng,ug,mg].}

\item{sink.unit}{The mass unit for the sink compartment [ng,ug,mg].}

\item{time.label.pos}{The position of the time label in the plots (see [graphics::legend()]).}

\item{mass.range}{The range of plottable mass (NULL for automatic min/max).}

\item{time.range}{The time range for the plots (NULL for automatic min/max).}

\item{...}{Additional parameter for plot directive.}
}
\value{
Filename of the produced movie.
}
\description{
Produces an overview movie for all mass-over time profiles for all compartments.
This will produce temporary images in a subdirectory (relative to the current working directory).
Naming of the subdirectory follows the template 'TAG_mass_movie_tmp'. Here, TAG is pulled from the tag of the supplied
DSkin Parameter Pack. If the directory does exist it will not be deleted or cleared before the image creation started.
}
\seealso{
\code{\link{dskin.parameter}}

\code{\link{dskin.simulate}}
}
