% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_plot.R
\name{dskin.movie.to.gif}
\alias{dskin.movie.to.gif}
\title{Converts a Movie to Gif}
\usage{
dskin.movie.to.gif(parameter, dskin.class = "mass", width = 500, fps = 25)
}
\arguments{
\item{parameter}{The DSkin Parameter Pack used to produce the source mp4 movie.}

\item{dskin.class}{The class of movie as a source [mass, conc].}

\item{width}{The desired width of the gif.}

\item{fps}{The desired FPS of the gif.}
}
\value{
Output filename of the produced gif.
}
\description{
Coverts a produced mp4 movie to gif file. The function needs ffmpeg in Path to work.
The output filename is constructed as 'TAG_masses.gif' for mass-over-time data or 'TAG_concs.gif' for concentration-depth profile data.
Here, TAG is the project tag defined in the DSkin Parameter Pack.
}
\seealso{
\code{\link{dskin.parameter}}
}
