% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.parameter}
\alias{dskin.parameter}
\title{Create a DSkin parameter pack}
\usage{
dskin.parameter(tag = "Unknown", n.layers = 1, layer.names = NULL,
  vehicle.name = "Donor", sink.name = "Sink", use.pk = TRUE)
}
\arguments{
\item{tag}{String tag to indicate a project name.}

\item{n.layers}{The number of barrier layers that shoudl be created.}

\item{layer.names}{Vector of layer names that should be the same length as n.layers. If NULL, default layer names will be used.}

\item{vehicle.name}{The name of the vehicle layer.}

\item{sink.name}{The name of the sink layer.}

\item{use.pk}{Logical value to indicate whether pharmacokinetic processes shoudl be simulated.}
}
\value{
The DSkin parameter pack.
}
\description{
Creates a DSkin parameter pack to configure the DSkin simulation run.
This data structure is the fundamental basis for every simulation and post-simulation analysis.
}
\examples{
p <- dskin.parameter("project_1", 2, c("SC", "DSL"), "TTS", "Blood", TRUE)
}
