% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_plot.R
\name{dskin.plot.concs}
\alias{dskin.plot.concs}
\title{Overview Concentration-depth Profile Plots}
\usage{
dskin.plot.concs(parameter, r.result = NULL, max.profiles = 5,
  plotting.times = NULL, path = NULL, max.cols = 2, layer.unit = "mg",
  conc.range = c(0, NULL), space.range = c(0, NULL), colors = c("red",
  "yellow", "springgreen", "royalblue"), ...)
}
\arguments{
\item{parameter}{The DSkin parameter pack.}

\item{r.result}{DSkin result data structure ('collect.data') by the R-version of DSkin. If NULL data is read from files defined by the Parameter Pack.}

\item{max.profiles}{The maximum number of profiles per plot (if not overridden by plotting.times).}

\item{plotting.times}{A vector of plotting times (time for every CDP). Overrides max.profiles if defined.}

\item{path}{The path to CDP files (NULL for wd).}

\item{max.cols}{The maximum number of columns in the plot.}

\item{layer.unit}{The mass unit for all layer compartments [ng,ug,mg].}

\item{conc.range}{The range of plottable concentration (NULL for automatic min/max).}

\item{space.range}{The space range for the plots (NULL for automatic min/max).}

\item{colors}{Vector of colors (or single color) for the color ramping procedure to colorize all curves in one subplot.}

\item{...}{Additional parameter for plot directive.}
}
\value{
None
}
\description{
Produces an overview plot for concentration-depth-profiles (CDP).
}
\seealso{
\code{\link{dskin.parameter}}

\code{\link{dskin.simulate}}
}
