% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_plot.R
\name{dskin.plot.masses}
\alias{dskin.plot.masses}
\title{Overview Mass-over-time Plots}
\usage{
dskin.plot.masses(parameter, r.result = NULL, path = NULL, max.cols = 3,
  plot.total = FALSE, total.main = "System", time.unit = "h",
  layer.unit = "mg", sink.unit = "ng", mass.range = c(0, NULL),
  time.range = c(0, NULL), ...)
}
\arguments{
\item{parameter}{The DSkin Parameter Pack.}

\item{r.result}{DSkin result data structure ('collect.data') by the R-version of DSkin. If NULL, data is read from produced files
defined by the Parameter Pack.}

\item{path}{The path to the mass-over-time files (NULL for wd).}

\item{max.cols}{The maximum number of columns in the plot.}

\item{plot.total}{Logical value to define if total system mass-over-time should be plotted.}

\item{total.main}{String of the system (plot.total) mass-over-time plot heading.}

\item{time.unit}{The unit of time used for the plot [sec,min,h].}

\item{layer.unit}{The mass unit for all layer compartments [ng,ug,mg].}

\item{sink.unit}{The mass unit for the sink compartment [ng,ug,mg].}

\item{mass.range}{The range of plottable mass (NULL for automatic min/max).}

\item{time.range}{The time range for the plots (NULL for automatic min/max).}

\item{...}{Additional parameter for plot directive.}
}
\value{
None
}
\description{
Produces an overview plot for all mass compartments and an optional overall system mass-over-time summary.
}
\seealso{
\code{\link{dskin.parameter}}

\code{\link{dskin.simulate}}
}
