% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_data.R
\name{dskin.read.concData}
\alias{dskin.read.concData}
\title{Read Concentration-depth Profiles}
\usage{
dskin.read.concData(parameter, path = NULL)
}
\arguments{
\item{parameter}{The DSkin parameter pack (see \code{\link{dskin.parameter}}).}

\item{path}{The path to the concentration-depth profile files (NULL for wd).}
}
\value{
A list with named compartments and DSkin class 'conc.data'.
}
\description{
Reads concentration-depth profiles from files produced by the DSkin Commandline Tool.
}
\seealso{
\code{\link{dskin.parameter}}
}
