% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_data.R
\name{dskin.read.massData}
\alias{dskin.read.massData}
\title{Read Mass-over-time Data}
\usage{
dskin.read.massData(parameter, path = NULL)
}
\arguments{
\item{parameter}{The DSkin parameter pack (see \code{\link{dskin.parameter}}).}

\item{path}{The path to the mass-over-time files (NULL for wd).}
}
\value{
Data frame with named compartments and DSkin class 'mass.data'.
}
\description{
Reads mass-over-time data from files produced by the DSkin Commandline Tool.
}
\seealso{
\code{\link{dskin.parameter}}
}
