% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_helper.R
\name{dskin.runtime.str}
\alias{dskin.runtime.str}
\title{DSkin Runtime String}
\usage{
dskin.runtime.str(result)
}
\arguments{
\item{result}{The DSkin result data structure (dskin class = 'collect.data').}
}
\value{
A human readable string of the runtime.
}
\description{
Produces a human readable string of the runtime of a DSkin simulation
}
\seealso{
\code{\link{dskin.simulate}}
}
