% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_bindings.R
\name{dskin.simulate}
\alias{dskin.simulate}
\title{Run DSkin Simulation}
\usage{
dskin.simulate(parameter, output = c("internal"))
}
\arguments{
\item{parameter}{The DSkin Parameter Pack or a list of DSkin Parameter Packs.}

\item{output}{Output options of the simulation.
Options are:
\describe{
 \item{internal}{Returns the results as an intenal R data structure}
 \item{files}{Produces output files in the current working direction}
}
The maximum number of output options are two and at least one option must be passed.}
}
\value{
DSkin internal R data strucure (base: list) with DSkin class 'collect.data' or named list (file_tag of the configurations as names)
of DSkin internal R data structures.
}
\description{
Runs a DSkin simulation based on a DSkin Parameter Pack.
If a list of configurations is used as input all computations will be run in parallel using the maximum available or needed compute cores
of the machine.
}
\seealso{
\code{\link{dskin.parameter}}
}
