% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.toJsonFile}
\alias{dskin.toJsonFile}
\title{Parameter Pack to JSON File}
\usage{
dskin.toJsonFile(parameter, file.name = "dskin_cfg.json")
}
\arguments{
\item{parameter}{The DSkin parameter pack.}

\item{file.name}{The name of the DSkin JSON file (DSkin configuration parameters) that can be processed by the DSkin Command Line Tool.}
}
\value{
None
}
\description{
Converts a dskin parameter pack to a JSON file.
}
\examples{
p <- dskin.parameter("project_1", 2, c("SC", "DSL"), "Patch", "Blood", TRUE)
dskin.toJsonFile(p, "new_cfg.json")
}
\seealso{
\code{\link{dskin.parameter}}
}
