% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dskin_params.R
\name{dskin.toJsonStr}
\alias{dskin.toJsonStr}
\title{Parameter Pack to JSON}
\usage{
dskin.toJsonStr(parameter)
}
\arguments{
\item{parameter}{The DSkin parameter pack (see \code{\link{dskin.parameter}}).}
}
\value{
The JSON representation of the DSkin parameters that can be processed by the DSkin Command Line Tool.
}
\description{
Converts a dskin parameter pack to a JSON string.
The string can be read by the DSkin Command Line Tool.
}
\examples{
p <- dskin.parameter("project_1", 2, c("SC", "DSL"), "Patch", "Blood", TRUE)
s <- dskin.toJsonStr(p)
}
\seealso{
\code{\link{dskin.parameter}}
}
